PR fixes.   - SarifToCsv support for Suppressions, Properties.   - Move IsEmptyEnumerable to ExtensionMethods.   - ResultLogJsonWriter writes properties in schema order as much as possible.   - Update Fortify unit tests for new Sarif ordering.
