+++ src/Samples/SarifToCsv/SarifToCsv.csproj
@@ -50,7 +50,9 @@
     <Compile Include="CsvWriter.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="SarifCsvColumnWriters.cs" />
     <Compile Include="SdkExtensions.cs" />
+    <Compile Include="WriteContext.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
