+++ src/Test.UnitTests.Sarif.Converters/TestData/FortifyFprConverter/ExpectedOutputs/TwoResultsWithNodeRefs.sarif
@@ -3,16 +3,6 @@
   "version": "2.1.0",
   "runs": [
     {
-      "automationDetails": {
-        "id": "The_Build_ID/"
-      },
-      "originalUriBaseIds": {
-        "SRCROOT": {
-          "uri": "file:///C:/projects/myproject/src/"
-        }
-      },
-      "language": "en-US",
-      "columnKind": "utf16CodeUnits",
       "results": [
         {
           "ruleIndex": 0,
@@ -244,6 +234,12 @@
           }
         }
       ],
+      "language": "en-US",
+      "originalUriBaseIds": {
+        "SRCROOT": {
+          "uri": "file:///C:/projects/myproject/src/"
+        }
+      },
       "artifacts": [
         {
           "location": {
@@ -273,6 +269,10 @@
           "encoding": "windows-1252"
         }
       ],
+      "automationDetails": {
+        "id": "The_Build_ID/"
+      },
+      "columnKind": "utf16CodeUnits",
       "threadFlowLocations": [
         {
           "location": {
@@ -412,6 +412,16 @@
             "return",
             "function"
           ]
+        }
+      ],
+      "taxonomies": [
+        {
+          "guid": "2B841697-D0DE-45DD-9F19-1EEE1312429",
+          "name": "CWE",
+          "organization": "MITRE",
+          "shortDescription": {
+            "text": "The MITRE Common Weakness Enumeration"
+          }
         }
       ]
     }
